+++ dtos/responses/event_test.go
@@ -34,7 +34,7 @@ func TestNewMultiEventsResponse(t *testing.T) {
 		{Id: "7a1707f0-166f-4c4b-bc9d-1d54c74e0137"},
 		{Id: "11111111-2222-3333-4444-555555555555"},
 	}
-	expectedTotalCount := uint32(len(expectedEvents))
+	expectedTotalCount := int64(len(expectedEvents))
 	actual := NewMultiEventsResponse(expectedRequestId, expectedMessage, expectedStatusCode, expectedTotalCount, expectedEvents)
 
 	assert.Equal(t, expectedRequestId, actual.RequestId)
